pyproject.toml
[project]
name = "illm"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0",
    "flask-sqlalchemy>=3.1",
    "flask-migrate>=4.0",
    "openai>=1.0",
    "python-dotenv>=1.0",
    "psycopg2-binary>=2.9",
    "pyyaml>=6.0",
    "authlib>=1.3",
    "requests>=2.31",
]

[dependency-groups]
dev = []
